requirements.txt
cachetools==5.3.1
flask==2.3.2
numpy==1.25.2
pydantic==2.1.1
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
pre-commit==3.3.3
alembic==1.11.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
Flask-SQLAlchemy==3.0.5
pgvector==0.2.1
Flask-Migrate==4.0.4
scikit-learn==1.3.0
supabase==1.0.4
pytesseract==0.3.10
pandas==2.0.3
redis==4.6.0
celery==5.3.1
sentence-transformers==2.2.2
azure-storage-blob==12.17.0

# Dev
pre-commit==3.3.3
flake8==6.1.0
autopep8==2.0.2
